Rename TokenListApiImpl file
