+++ src/api/tokenList/TokenListApiMock.ts
@@ -1,5 +1,5 @@
 import { TokenDetails } from 'types'
-import { TokenList } from './TokenListApiImpl'
+import { TokenList } from './TokenListApi'
 export class TokenListApiMock implements TokenList {
   private _tokenList: TokenDetails[]
