+++ src/services/factories/getTokenFromExchange.ts
@@ -6,7 +6,7 @@ import { getImageUrl, log, getToken } from 'utils'
 import { getErc20Info } from '../helpers'
 import { Erc20Api } from 'api/erc20/Erc20Api'
 import { ExchangeApi } from 'api/exchange/ExchangeApi'
-import { TokenList } from 'api/tokenList/TokenListApiImpl'
+import { TokenList } from 'api/tokenList/TokenListApi'
 
 interface TokenFromErc20Params {
   tokenAddress: string
